Updated Enable-LocationServices Function
Added the ability for the script to check to see if the 'Status' DWORD exists under HKLM:\SYSTEM\CurrentControlSet\Services\lfsvc\Service\Configuration
If it exists, it'll check the value. If it doesn't exist, it'll create it with the correct value. Previously, it didn't exist on new Windows 10 Machines and the script would fail since it never created it and only checked for the value.